requirements.txt
Django==3.1.3
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
django-cors-headers==3.5.0
Markdown==3.3.3
phonenumbers==8.12.13
pycparser==2.20
pytz==2020.4
sqlparse==0.4.1
dj-database-url==0.5.0
psycopg2==2.8.2
Pillow==8.0.1
whitenoise==5.2.0
gunicorn==20.0.4